Ensure buffers are never exhausted when looping
If we're near the end of a file, it was possible for the
initial read to only fill the buffer with a few bytes, which
was not enough for continuous playback. When the buffer was
exhausted, OpenAL puts the source into a AL_STOPPED state,
and causes looped music to be unintentionally interrupted.

I've avoided this case by continuing to fill the buffer after
rewinding the file, so that each buffer is always full.